Create workspace. Move Constant into its own file.
