+++ .gitignore
@@ -500,3 +500,20 @@ tools/standalone_visualizer/release-builds/*
 gcloud_config.json
 
 *.bundle.js
+
+# Xcode things to ignore
+## Various settings
+*.pbxuser
+!default.pbxuser
+*.mode1v3
+!default.mode1v3
+*.mode2v3
+!default.mode2v3
+*.perspectivev3
+!default.perspectivev3
+xcuserdata/
+
+## Other
+*.moved-aside
+*.xccheckout
+*.xcscmblueprint
